iploadbalancing: mv all types in specific file
