+++ ovh/resource_ovh_iploadbalancing_http_farm_server.go
@@ -9,23 +9,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
 )
 
-type IpLoadbalancingHttpFarmServer struct {
-	BackendId            int     `json:"backendId,omitempty"`
-	ServerId             int     `json:"serverId,omitempty"`
-	FarmId               int     `json:"farmId,omitempty"`
-	DisplayName          *string `json:"displayName,omitempty"`
-	Address              string  `json:"address"`
-	Cookie               *string `json:"cookie,omitempty"`
-	Port                 *int    `json:"port"`
-	ProxyProtocolVersion *string `json:"proxyProtocolVersion"`
-	Chain                *string `json:"chain"`
-	Weight               *int    `json:"weight"`
-	Probe                *bool   `json:"probe"`
-	Ssl                  *bool   `json:"ssl"`
-	Backup               *bool   `json:"backup"`
-	Status               string  `json:"status"`
-}
-
 func resourceIpLoadbalancingHttpFarmServer() *schema.Resource {
 	return &schema.Resource{
 		Create: resourceIpLoadbalancingHttpFarmServerCreate,
